debugger/debugger_ia/requirements.txt
python-dotenv
google-generativeai
flet